move User to a specific model file
